Rename ShortGraphQLService into shortGraphQL (#862)
