+++ frontend/src/dep.ts
@@ -20,8 +20,8 @@ import { ShortHTTPApi } from './service/ShortHTTP.api';
 import { DynamicDecisionService } from './service/feature-decision/DynamicDecision.service';
 import { ShortLinkService } from './service/ShortLink.service';
 import { AnalyticsService } from './service/Analytics.service';
-import { ChangeLogGraphQLApi } from './service/ShortGraphQLService/ChangeLogGraphQL.api';
-import { ShortLinkGraphQLApi } from './service/ShortGraphQLService/ShortLinkGraphQL.api';
+import { ChangeLogGraphQLApi } from './service/shortGraphQL/ChangeLogGraphQL.api';
+import { ShortLinkGraphQLApi } from './service/shortGraphQL/ShortLinkGraphQL.api';
 
 export function initEnvService(): EnvService {
   return new EnvService();
