+++ frontend/src/service/shortGraphQL/ShortLinkGraphQL.api.ts
