+++ frontend/src/service/shortGraphQL/schema.ts
